+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to the Imperative package will be documented in this file.
 
 ## Recent Changes
 
-- BugFix: Clean up uses of execSync in Imperative where it makes sense to do so.
+- BugFix: Updated plugins `--login` command option to behave as expected when running in an NPM 9 environment
+- BugFix: Cleaned up uses of execSync in Imperative where it makes sense to do so.
 
 ## `5.7.0`