+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the Imperative package will be documented in this file.
 
+## Recent Changes
+
+- BugFix: Updated plugins `--login` command option to behave as expected when running in an NPM 9 environment
+- BugFix: Cleaned up uses of execSync in Imperative where it makes sense to do so.
+
 ## `5.7.0`
 
 - Enhancement: Add `zowe config report-env` command to show a diagnostic report of the CLI's working environment.